requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dominate==2.8.0
Flask==2.2.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.4
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.1
requests==2.31.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.2
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
pip==23.1.2
psycopg2-binary